[backend] Enhance garbage collection everywhere

+++ openbas-api/src/main/java/io/openbas/injectors/caldera/service/CalderaGarbageCollectorService.java
@@ -19,6 +19,7 @@ import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
@@ -58,34 +59,24 @@ public class CalderaGarbageCollectorService implements Runnable {
         log.info("Running Caldera injector garbage collector...");
         List<Endpoint> endpoints = this.endpointService.endpoints(EndpointSpecification.findEndpointsForExecution());
         endpoints.forEach(endpoint -> {
-            if ((now().toEpochMilli() - endpoint.getCreatedAt().toEpochMilli()) > KILL_TTL) {
-                try {
-                    client.killAgent(endpoint);
-                } catch (RuntimeException e) {
-                    log.info("Failed to kill agent, probably already killed");
-                }
-            }
             if ((now().toEpochMilli() - endpoint.getCreatedAt().toEpochMilli()) > DELETE_TTL) {
                 this.endpointService.deleteEndpoint(endpoint.getId());
-                try {
-                    client.deleteAgent(endpoint);
-                } catch (RuntimeException e) {
-                    log.severe("Failed to delete agent");
-                }
             }
         });
         try {
             List<Agent> agents = this.client.agents();
+            List<String> killedAgents = new ArrayList<>();
             agents.forEach(agent -> {
-                if (agent.getExe_name().contains("executor") && (now().toEpochMilli() - Time.toInstant(agent.getCreated()).toEpochMilli()) > KILL_TTL) {
+                if (agent.getExe_name().contains("executor") && (now().toEpochMilli() - Time.toInstant(agent.getCreated()).toEpochMilli()) > KILL_TTL && (now().toEpochMilli() - Time.toInstant(agent.getLast_seen()).toEpochMilli()) < KILL_TTL) {
                     try {
                         log.info("Killing agent " + agent.getHost());
                         client.killAgent(agent);
+                        killedAgents.add(agent.getPaw());
                     } catch (RuntimeException e) {
                         log.info("Failed to kill agent, probably already killed");
                     }
                 }
-                if (agent.getExe_name().contains("executor") && (now().toEpochMilli() - Time.toInstant(agent.getCreated()).toEpochMilli()) > DELETE_TTL) {
+                if (agent.getExe_name().contains("executor") && (now().toEpochMilli() - Time.toInstant(agent.getCreated()).toEpochMilli()) > DELETE_TTL && !killedAgents.contains(agent.getPaw())) {
                     try {
                         log.info("Deleting agent " + agent.getHost());
                         client.deleteAgent(agent);
